Rename mcdocAttributes file in nbt package
