+++ packages/nbt/src/mcdocAttributes.ts
